pyproject.toml
[project]
name = "safecity-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.75.0",
    "duckdb>=1.4.3",
    "folium>=0.20.0",
    "jupyter>=1.1.1",
    "litellm>=1.80.10",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "streamlit-folium>=0.25.3",
]
